api/requirements.txt
flask
requests
mysql.connector.python
sqlalchemy
pandas
kafka-python
confluent-kafka
kubernetes
